Connecting database & verifying username/passsword
Is partially storing the username in the cookie along with a bunch of
garbage information.

Thus far I've tried doing it via server session data and several other
methods but I'm beginning to think this is to do with the fact the
server is on devweb more than anything as the cookie seems to be storing
itself fine, might just have to scrape the cookie just after creation of
this junk data but it'd 3.35 and things are getting blurry.