mainly done the category page.
Populated the database with mens stuff, using the DBfiller and phpmyadmin.

This enabled me to finish the category page. It now takes in args and should display items that match that category. Done the php and sass for this page along with podering about how much I hate databases

Created one database connection file that connects t the database this means that the login details are only in one place. Unfortunatly the login.php page connects the to DB in the weirdest way in hitory and I couldn't get it working. So could @epailes get it working the same way.