Basic Cart functionality again
Items can be added to cart, if the user tries to access cart with
nothing in it they are redirected to the empty-cart page instead of
getting a string displayed.

Remove button is not functional yet, start of the code is in place for
it.
Fixed bug where items load more than once in the basket/price is doubled
in total basket value.