Got add to cart working.
This involved rewriting a vast majority of the current code and refactoring anything that was keep.

The product information stored in the cookie used to just store the product name, when we needed further information, such as size and colour. Products are now being stored in the cookie in an array. This array contains the product arrays which contains the information about the users cart added products.